Mk 1 Mod 1 Version of the MacDavies Suppressor
This version changes the baffle design, adding more baffles and staggering the arrangement of turbulence holes. The holes are now two to a baffle instead of four, and alternate between the XZ and XY axis per baffle. They also decrease by 0.5mm in radius with each baffle till they are 2mm in radius. This arrangement should do better and creating turbulence within the expansion volumes and decelerating the gas across the bore.